+++ simulation-system/libs/csle-rest-api/tests/test_resources_users.py
@@ -8,7 +8,6 @@ from csle_common.dao.emulation_config.cluster_config import ClusterConfig
 from csle_common.dao.emulation_config.cluster_node import ClusterNode
 from csle_common.dao.emulation_config.config import Config
 from csle_common.dao.management.management_user import ManagementUser
-from flask import jsonify
 
 import csle_rest_api.constants.constants as api_constants
 from csle_rest_api.rest_api import create_app
@@ -29,65 +28,6 @@ class TestResourcesUsersSuite(object):
         return create_app(
             static_folder="../../../../../management-system/csle-mgmt-webapp/build"
         )
-
-    @pytest.fixture
-    def logged_in_as_admin(self, mocker):
-        """
-        Fixture for mocking the logged-in-as-admin side effect
-
-        :param mocker: the pytest mocker object
-        :return: the logged in as admin fixture for mocking the logged in check
-        """
-        def check_if_user_is_authorized(request, requires_admin):
-            return None
-
-        check_if_user_is_authorized_mock = mocker.MagicMock(
-            side_effect=check_if_user_is_authorized
-        )
-        return check_if_user_is_authorized_mock
-
-    @pytest.fixture
-    def logged_in(self, mocker):
-        """
-        Fixture for mocking the logged-in side effect
-
-        :param mocker: the pytest mocker object
-        :return: the logged in fixture for mocking the logged in check
-        """
-        def check_if_user_is_authorized(request, requires_admin):
-            if requires_admin:
-                response = jsonify({})
-                response.headers.add(
-                    api_constants.MGMT_WEBAPP.ACCESS_CONTROL_ALLOW_ORIGIN_HEADER, "*"
-                )
-                return response, constants.HTTPS.UNAUTHORIZED_STATUS_CODE
-            return None
-
-        check_if_user_is_authorized_mock = mocker.MagicMock(
-            side_effect=check_if_user_is_authorized
-        )
-        return check_if_user_is_authorized_mock
-
-    @pytest.fixture
-    def not_logged_in(self, mocker):
-        """
-        Fixture for mocking the not-logged-in side effect
-
-        :param mocker: the pytest mocker object
-        :return: the not-logged-in fixture for mocking the logged in check
-        """
-
-        def check_if_user_is_authorized(request, requires_admin):
-            response = jsonify({})
-            response.headers.add(
-                api_constants.MGMT_WEBAPP.ACCESS_CONTROL_ALLOW_ORIGIN_HEADER, "*"
-            )
-            return response, constants.HTTPS.UNAUTHORIZED_STATUS_CODE
-
-        check_if_user_is_authorized_mock = mocker.MagicMock(
-            side_effect=check_if_user_is_authorized
-        )
-        return check_if_user_is_authorized_mock
 
     @pytest.fixture
     def management_users(self, mocker):
@@ -230,50 +170,7 @@ class TestResourcesUsersSuite(object):
         save_management_user_mock = mocker.MagicMock(side_effect=save_management_user)
         return save_management_user_mock
 
-    @pytest.fixture
-    def authorized(self, mocker):
-        """
-        Fixture for mocking the check_if_user_edit_is_authorized function
-
-        :param mocker: the pytest mocker object
-        :return: fixture for mocking the check_if_user_edit_is_authorized function
-        """
-        def check_if_user_edit_is_authorized(request, user: ManagementUser):
-            token = request.args.get(api_constants.MGMT_WEBAPP.TOKEN_QUERY_PARAM)
-            if token != "":
-                return user
-            response = jsonify({})
-            response.headers.add(
-                api_constants.MGMT_WEBAPP.ACCESS_CONTROL_ALLOW_ORIGIN_HEADER, "*"
-            )
-            return response, constants.HTTPS.UNAUTHORIZED_STATUS_CODE
-        check_if_user_edit_is_authorized_mock = mocker.MagicMock(
-            side_effect=check_if_user_edit_is_authorized
-        )
-        return check_if_user_edit_is_authorized_mock
-
-    @pytest.fixture
-    def unauthorized(self, mocker):
-        """
-        Fixture for mocking the check_if_user_edit_is_authorized function
-
-        :param mocker: the pytest mocker object
-        :return: fixture for mocking the check_if_user_edit_is_authorized function
-        """
-        def check_if_user_edit_is_authorized(request, user: ManagementUser):
-            token = request.args.get(api_constants.MGMT_WEBAPP.TOKEN_QUERY_PARAM)
-            if token == "":
-                response = jsonify({})
-                response.headers.add(
-                    api_constants.MGMT_WEBAPP.ACCESS_CONTROL_ALLOW_ORIGIN_HEADER, "*"
-                )
-                return response, constants.HTTPS.UNAUTHORIZED_STATUS_CODE
-            return user
-        check_if_user_edit_is_authorized_mock = mocker.MagicMock(
-            side_effect=check_if_user_edit_is_authorized
-        )
-        return check_if_user_edit_is_authorized_mock
-
+    @pytest.mark.usefixtures("authorized", "unauthorized", "logged_in", "logged_in_as_admin", "not_logged_in")
     def test_users(
             self,
             flask_app,
